LayoutTests/fast/images/text-recognition/ios/do-not-coalesce-non-adjacent-selection-rects-in-image-overlay-expected.txt

PASS selectionRectsAfterSelectingAll.length is 2
PASS successfullyParsed is true

TEST COMPLETE
